+++ packages/sdk-redux-examples/sdk-redux-react-typecript/src/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import App from "./App";
 import { store } from "./redux/store";
 import { Provider } from "react-redux";
@@ -8,21 +8,26 @@ import { CssBaseline, ThemeProvider } from "@mui/material";
 import { WagmiConfig } from "wagmi";
 import { chains, wagmiClient } from "./wagmiAndRainbowKit";
 import { RainbowKitProvider } from "@rainbow-me/rainbowkit";
-import '@rainbow-me/rainbowkit/styles.css';
+import "@rainbow-me/rainbowkit/styles.css";
 
-ReactDOM.render(
-    <React.StrictMode>
-        <ThemeProvider theme={theme}>
-            {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
-            <CssBaseline />
-            <Provider store={store}>
-                <WagmiConfig client={wagmiClient}>
-                    <RainbowKitProvider chains={chains}>
-                        <App />
-                    </RainbowKitProvider>
-                </WagmiConfig>
-            </Provider>
-        </ThemeProvider>
-    </React.StrictMode>,
-    document.getElementById("root")
-);
+const root = ReactDOM.createRoot(document.getElementById("root")!);
+
+const Root = () => {
+    return (
+        <React.StrictMode>
+            <ThemeProvider theme={theme}>
+                {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
+                <CssBaseline />
+                <Provider store={store}>
+                    <WagmiConfig client={wagmiClient}>
+                        <RainbowKitProvider chains={chains}>
+                            <App />
+                        </RainbowKitProvider>
+                    </WagmiConfig>
+                </Provider>
+            </ThemeProvider>
+        </React.StrictMode>
+    );
+};
+
+root.render(<Root />);